Рефакторинг. Дабавил метод копирования файла с прогрессом.
