+++ 7.AspnetMvc/Lesson-7/FileCommander.Interfaces/IFileService.cs
@@ -1,0 +1,7 @@
+﻿namespace FileCommander.Interfaces;
+
+public interface IFileService
+{
+    Task CopyFileAsync(string source, string destination, IProgress<int>? progress = null, CancellationToken cancel = default);
+    void CreateReport(string reportFileName, string sourceFileName);
+}
