+++ 7.AspnetMvc/Lesson-7/FileCommander.Interfaces/Infrastructure/IEventBus.cs
@@ -1,0 +1,8 @@
+namespace FileCommander.Interfaces.Infrastructure;
+
+public interface IEventBus
+{
+    public void Publish(IIntegrationEvent @event);
+    public void Subscribe<TEvent>(Action<IIntegrationEvent> handler)
+        where TEvent : IIntegrationEvent;
+}
