+++ 7.AspnetMvc/Lesson-7/FileCommander.Interfaces/Infrastructure/IIntegrationEvent.cs
@@ -1,0 +1,6 @@
+namespace FileCommander.Interfaces.Infrastructure;
+
+public interface IIntegrationEvent
+{
+    public string Name { get; set; }
+}
