+++ 7.AspnetMvc/Lesson-7/FileCommander.Interfaces/Reports/ICreateReportOptions.cs
@@ -1,0 +1,7 @@
+﻿namespace FileCommander.Interfaces.Reports
+{
+    public interface ICreateReportOptions
+    {
+        bool OpenAfterCreate { get; init; }
+    }
+}
