+++ 7.AspnetMvc/Lesson-7/FileCommander.Interfaces/Reports/IReport.cs
@@ -1,0 +1,6 @@
+namespace FileCommander.Interfaces.Reports;
+
+public interface IReport
+{
+    void MakeReport(string reportFileName, string sourceFileName, ICreateReportOptions options);
+}
