+++ 7.AspnetMvc/Lesson-7/FileCommander.Interfaces/Reports/IReportResolver.cs
@@ -1,0 +1,6 @@
+namespace FileCommander.Interfaces.Reports;
+
+public interface IReportResolver
+{
+    IReport GetReportByFileName(string fileName);
+}
