+++ 7.AspnetMvc/Lesson-7/FileCommander.Services/ReportResolver.cs
@@ -1,0 +1,19 @@
+using Autofac.Features.Indexed;
+using FileCommander.Interfaces.Reports;
+
+namespace FileCommander.Services;
+
+public class ReportResolver : IReportResolver
+{
+    private readonly IIndex<string, IReport> _reports;
+
+    public ReportResolver(IIndex<string, IReport> reports) => _reports = reports;
+
+    public IReport GetReportByFileName(string fileName)
+    {
+        if (Directory.Exists(fileName))
+            return _reports["directory"];
+
+        return _reports["file"];
+    }
+}
