+++ 7.AspnetMvc/Lesson-7/FileCommander.Services/Reports/CreateReportOptions.cs
@@ -1,0 +1,8 @@
+﻿using FileCommander.Interfaces.Reports;
+
+namespace FileCommander.Services.Reports;
+
+public class CreateReportOptions : ICreateReportOptions
+{
+    public bool OpenAfterCreate { get; init; } = false;
+}
