+++ 7.AspnetMvc/Lesson-7/FileCommander/App.xaml.cs
@@ -1,17 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Configuration;
-using System.Data;
-using System.Linq;
-using System.Threading.Tasks;
 using System.Windows;
 using Autofac;
-using Autofac.Core;
 using Autofac.Extensions.DependencyInjection;
-using FileCommander.Reports;
-using FileCommander.Reports.Interfaces;
+using FileCommander.Interfaces;
+using FileCommander.Interfaces.Reports;
 using FileCommander.Services;
-using FileCommander.Services.Interfaces;
+using FileCommander.Services.Reports;
 using FileCommander.ViewModels;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
