--- 7.AspnetMvc/Lesson-7/FileCommander/Reports/CreateReportOptions.cs
@@ -1,6 +1,0 @@
-﻿namespace FileCommander.Reports;
-
-public class CreateReportOptions
-{
-    public bool OpenAfterCreate { get; init;} = false;
-}
