--- 7.AspnetMvc/Lesson-7/FileCommander/Reports/Interfaces/IReport.cs
@@ -1,6 +1,0 @@
-namespace FileCommander.Reports.Interfaces;
-
-public interface IReport
-{
-    void MakeReport(string reportFileName, string sourceFileName, CreateReportOptions options);
-}
