--- 7.AspnetMvc/Lesson-7/FileCommander/Reports/Interfaces/IReportResolver.cs
@@ -1,6 +1,0 @@
-namespace FileCommander.Reports.Interfaces;
-
-public interface IReportResolver
-{
-    IReport GetReportByFileName(string fileName);
-}
