--- 7.AspnetMvc/Lesson-7/FileCommander/Services/Interfaces/IFileService.cs
@@ -1,7 +1,0 @@
-﻿namespace FileCommander.Services.Interfaces
-{
-    public interface IFileService
-    {
-        void CreateReport(string reportFileName, string sourceFileName);
-    }
-}
