+++ 7.AspnetMvc/Lesson-7/Infrastructure/EventArguments/ProgressArgs.cs
@@ -1,0 +1,8 @@
+﻿namespace FileCommander.Infrastructure.EventArguments;
+
+public class ProgressArgs : EventArgs
+{
+    public int Progress { get; }
+
+    public ProgressArgs(int progress) => Progress = progress;
+}
