+++ 7.AspnetMvc/Lesson-7/Infrastructure/EventBus/Events/IntegrationEvent.cs
@@ -1,0 +1,8 @@
+using FileCommander.Interfaces.Infrastructure;
+
+namespace FileCommander.Infrastructure.EventBus.Events;
+
+public abstract class IntegrationEvent : IIntegrationEvent
+{
+    public string Name { get; set; }
+}
