+++ 7.AspnetMvc/Lesson-7/Infrastructure/FileCommander.Infrastructure.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\FileCommander.Interfaces\FileCommander.Interfaces.csproj" />
+  </ItemGroup>
+
+</Project>
